NetApp: E-Series driver using invalid host-types
The NetApp E-Series driver is using a dict of hard-coded host-type
values, some of which are no longer valid. There is a list of updated
host-types that are valid, but they must be validated against what
the storage array reports as supported.

This patch removes invalid or unnecessary host-types from library.py,
and adds a verification using the host-types API to ensure that the
host-type selected by the user is valid for the storage-array.

Closes-Bug: #1555946
Change-Id: I94090eb6b2d9c7965577a462f5e70033926fa27d