Update extend_volume API to use versionedobjects
The following patch updates extend_volume API to use
volume versionedobjects.  Changes were made to be
backwards compatible with older RPC clients.  It only
includes changes to the core cinder code.  Changes in
the drivers are left to each driver maintainer to
update.

Note that this patch DOES NOT try to use object dot
notation everywhere, since it would increase the
size of the patch.  Instead, it will be done in
subsequent patches.

Change-Id: I4192b35f8f3e2cd51b94db10b44727f23592ea86
Partial-Implements: blueprint cinder-objects