NetApp: Cleanup if E-Series volume create fails
The creation of a volume may fail in such a manner
that a volume in a partial state has been created.
When Openstack comes back re-attempting the creation
operation it fails because the partial volume with
the name already exists.

Solution: When there is a failure creating a volume,
check if a partial volume was created and delete it
so any retry will be successful and the storage device
does not contain any dead volumes.

Closes-Bug: 1506940
Change-Id: Id93527a4cab314f27a13d8c08b772fed538d1092