VMware: Relocate volume only during no disk space
During volume extend, we try to relocate the volume to a
different datastore if there is an error during vCenter
extend API call. But we need to relocate the volume to a
different datastore only if there is no space in the
current datastore to accommodate the extended volume. This
patch fixes the issue by relocating the volume only if the
exception raised is NoDiskSpaceException. It also refactors
the code to remove a redundant helper method for extending
virtual disk.

Closes-Bug: #1506422
Change-Id: Ic64eb697b49ffb3a435ffc4d969b5b7f79af91c6