Case sensitivity problem in cinder scheduler
Volume creation fails when we use backend name which consists of
same characters as the name used before but different upper/lower case.
It happens in MySQL environment.

e.g.
1. configure 'lvm' as backend name and start cinder
2. reconfigure 'LVM' as backend name and restart cinder

It is caused by the difference of handling of case sensitivity
for backend name.

When checking the existence of cinder-volume/backend to judge
whether new cinder-volume service row in the DB should be created
or not, the search-logic is case insensitive for backend name.

On the other hand, when searching backend in create_volume processing,
the search-logic in cinder-scheduler is case sensitive for backend name.

That is why I change the former search-logic
from case insensitive one to case sensitive one.

Closes-Bug: #1496694
Change-Id: Ibb1565a183961031495b73e985499455a1ea4c7b