Zfssaiscsi driver should not use 'default' initiator group
The zfssaiscsi driver falls back to using the 'default'
initiator group when it fails to find a valid initiator group
configured. This causes all LUNs attached to instances on that
server to be presented to every server using the target group.

This fix makes the driver use the 'default' group only if
configured in the cinder.conf file. It also throws an error when
initiator group is given without any initiators.

DocImpact
Change-Id: I417688cf632baa727e5aa339c8d362761a38dcc8
Closes-Bug: #1508548