Check backup service before backup delete
Right now, if a user tries to delete a backup and the cinder-backup
service is down, the backup status will be error and no message will
be provided to the user. This patch will add a check to see that the
service is running before issuing the rpcapi cast to delete the backup,
and inform the user that the service is down if the check fails. However,
it is still possible with this fix for the service to go down during the
time between the check and when the backup is actually deleted.

Change-Id: Ia9914bfaae8ed9b6972953fc1e78c4471fa1661a
Closes-bug: #1321519