EMC VMAX - Change naming convention for MV and SG for FAST
Up until now we assumed a one to one relationship between FAST policy
and Pool. This does not have to be the case. A FAST policy can span
multiple pools. To ensure uniqueness we will change the SG nand MV
name to contain the FAST policy and not the pool name.

Change-Id: I4d7cc4f8922d08b886b8eb3eb9b1a2b3f4340cc1
Closes-Bug: #1515181