Storwize: Add force flag on vdisk mapping call
If a vdisk is already mapped and multihostmap is True,
use the force flag on the first call to mkvdiskhostmap.
There is no functional change associated with this defect
for this method, a vdisk that is already mapped to a
host without the multihostmap flag will fail and it
will be successful if the multihostmap is True.

The code was waiting for the first call to svcinfo
mkvdiskhostmap to fail before trying a second time with
the force option if multihostmap is set to true. This fix
adds the force option before the call to svcinfo
mkvdiskhostmap to avoid this overhead. The svcinfo
mkvdiskhostmap call accepts the force option when
a volume is not already mapped, and mkvdiskhostmap()
already has a force parameter,
so the existing testcase is still valid.

Fixed a bug where mkvdiskhostmap() was returning
data, and add a testcase to check that it is not
returning data.

Change-Id: I2320aeec9422625ce2a01aa5d28801207e8671a5
Closes-Bug: 1511751