Cleanup orphaned code from brick LVM
There were several methods in the brick lvm code that were no longer
being used anywhere. While some of these have potential for being useful,
they are not in use and should not be sitting around the codebase waiting
for someone to need them.

This patch removes unused methods as well as their associated unit tests.

Change-Id: Ib3aba7b6b51a525855c5357e4164478f26e9ee98