Use the catalog to retrieve glance_api_servers
Currently the config option glance_api_servers must be
set to give the correct list of glance API URLs. This
patch enables the use of the catalog to find the glance
endpoint when this config option is set to None, and
changes its default to None as well.

Closes-bug: #1371644
DocImpact: The config options glance_host and glance_port
have been removed, the default value for glance_api_servers
has been changed to None, and a new config option
glance_catalog_info has been added to glance_opts.
Change-Id: I261da528eb008f6b2520aef043c06697bd5ce352