VNX: Fix issue in deleting cg/cgsnapshot
If user deletes a cg/cgsnapshot quickly after its creation,
driver would report 'not found' warning, and return success
for the deletion, but actually, the object underlying the VNX
is not deleted.

This fix tries to wait the cg and cg snapshot to be available
to make sure it's eligible for deletion.

liberty-backport-potential
Closes-Bug: 1499615
(cherry picked from commit f7b5655d188bfde96333760bbe634f45d4a16395)

Change-Id: Ifae57b9e95e01b1789a37ac7c03e9aad65cd50f7