VMware: Fix volume copy across vCenter datacenters
Currently volume creation from another volume or snapshot fails
if the source volume/snapshot and the destination volume are in
different vCenter datacenters. This is because we are always
passing the source datacenter folder as the location of the
destination volume while calling vCenter clone API. This patch
fixes the problem by passing the correct datacenter folder while
copying volume or snapshot.

Change-Id: I3bfbc149fc75fbff303b9c9b1c833b499fa2814e
Closes-bug: #1521991