Update manage_existing to use volume object
The following patch updates manage_existing APIs to use volume versioned
objects. Changes were made to be backward compatible with older RPC
clients. Changes in the drivers are left to each driver maintainer to
update.

Note that this patch DOES NOT try to use object dot notation everywhere
to keep the size small. This cleanup will be done in the future.

Change-Id: Ie806336ad95834b091c5942d7dfa13345006c1c8
Partial-Implements: blueprint cinder-objects