Add test for checking object compatibilities
oslo.versionedobjects has a fixture test that can check object
compatibilities with previous versions of the object. It isn't a
catch-all for making sure everything works between the versions, but it
will at least check to make sure obj_to_primitive() won't totally blow
up when trying to backport to a previous version.

Change-Id: I2d3c26357f8ed9104bb9a936273f9cb8031715ca