Remove name_id when creating volume with cache enabled
When image_volume_cache_enabled = True, Cinder will create
a cached image volume. But when creating this image volume,
the name_id is set to id of newly created volume, which causes
the image volume is unable to be created in the underlying storage
system.

This fix will remove the name_id so that volume.name can refer to
correct LUN on storage system

Also, a minor fix: passing volume.id to delete_volume since
it only accept volume id.

Change-Id: Ib47b189a525d8a9bd9026cc20f62542aede9c9d4
Closes-bug: 1528087