Wrap the method to calculate virtual free capacity
There is a logic already there to calculate virtual
free capacity in cinder/scheduler/weights/capacity.py.
This patch is just to wrap the same logic into a
method in case to avoid the inconsistency when there
are some other code paths to use the the same logic
seperately. The reason for the patch is that:
the patch for blueprint capacity-headroom has the
same code path, and some reviewers prefer to use a
single method to call them in the same way.

Partial-Implements: blueprint capacity-headroom
Change-Id: I7554699760f2173655cb4fae434bd4933be77c9c