Remove QoS settings from SolidFire attributes
There's a limit in the number of parent key objects that
are stored on the SolidFire cluster object attributes (10).

While we're not hitting the limit yet, we're getting dangerously
close, so this patch removes the unnecessary QoS keys that are
currently added.  These keys can safely be removed because they
are stored by Cinder already and can be retrieved via the type
setting, or in the case of retype, the retype info that's stored
on the volume attribute.

Other keys shouldn't be modified at this time because they are
used internally by the driver to make various decisions, and
turning them into sub-keys or removing them altogether will
involve more far reaching changes that aren't necessary in the
short term to fix this particular bug.

Change-Id: I617ba41331e95c45f86a08440e24e6c407033a46
Closes-Bug: #1530212