Don’t log warnings for image cache when disabled
Right now we do a check for the internal tenant and log a warning before
we check to see if the cache is even turned on for that backend. We
should be only checking for the internal tenant if the image cache has
been configured for the backend.

In addition we shouldn’t log a warning for this since there is already
a warning for the internal context when we try and create it if it is
not configured correctly… to reduce the warning spam this change
switches the log message to an info level.

Change-Id: Iddc4648543d268a4b3b8d557c94c4d2876654e22
Closes-Bug: #1530964