Delete: Initial framework
Added deleted column to backup db

find all backups who are dependent on given backup.
If any of the dependent backup is not marked for deletion,
then just mark backup to be deleted and return.
If all of the dependent backups are marked for deletion,
then delete all the backups which can be.

Signed-off-by: shishir gowda <shishir.gowda@ril.com>