Remove extra get_active_zoneset query from Brocade CLI
A recent brocade zoning driver patch for bug-1465207 has introduced
an extra query for the "active_zone_set" from the switch for
every add_zone operation.

Performance can be optimized by comparing the "active_zone_set",
already passed in, with the zones to be updated instead. Removing
the need to re-query for the "active_zone_set" to determine if
the "active_cfg" requires recreation.

Closes-bug #1532272
Change-Id: Ic9a7ce8c3f4bbaddccf42e9640eab4750acce484