VMware: Support for non-default port
Currently the VMDK driver fails to connect to vCenter server
if vCenter is configured to listen to any port other than the
default port 443. This patch fixes it by adding support for a
new config option 'vmware_host_port' which specifies the port
number to connect to vCenter server.

DocImpact
  Added config option 'vmware_host_port' which specifies the
  port number to connect to vCenter server.

Change-Id: I71846cdc09bc1ff2a7e35d45aa7c639c468f1418
Co-Authored-By: Johnson koil raj <johnson.raj@hp.com>
Closes-Bug: #1387074