support user given display-name
in ceph we create snapshots as "backup.<backup-id>.snap.<time-stamp>
time-stamp is not the same as created at, so we store this in a new
column of the backup db. Initially, we used to store this string
in the display_name field

Additionally, we now are not deleting backups on src volumes to
support deletion of backups. We now need to identify the right
latest backup (not necessarily the latest if it was deleted)

Signed-off-by: shishir gowda <shishir.gowda@ril.com>