deletion: remove snapshot from src too
on actual deletion(not on mark for deletion), we need clear
out the snapshot from the src pool too.

misc fixes

Save current time_stamp. Ceph truncates timestamp to 0.2f
we need to save the same so that we can access the same
in src and DSS

Connect to ceph pool and delete rbd snap too on deletion
from DSS

Signed-off-by: shishir gowda <shishir.gowda@ril.com>