Huawei: Record and check LUN wwn
Currently we use LUN ID as the only identifier for a LUN,
this will cause problems. For example:
User create a volume in cinder, the LUN ID on array is 10.
Now user delete the LUN directly on the array by mistaken
and create a new LUN on the array(no matter create it directly
on array or on the cinder side), the LUN ID of the newly
created LUN will be 10 again! Then every operation on the very
early created volume on the cinder side will be done on the
newly created LUN!

Closes-Bug: #1535590
Change-Id: If671bf6c0ba342e2028d26aa53dc5ae09c730130
(cherry picked from commit f425bd5697398f88b1a046357016f92039bb5333)