Support ZeroMQ messaging driver in cinder
NOTE:This patch introduces support for ZeroMQ driver for cinder single
backend case.
Multi-backend will be addressed in the next patch as part of the
blueprint.

CHANGES:ZeroMQ driver requires hostname for message delivery as there is no
broker inbetween.
So, extract the hostname and feed to the messaging client for zeromq
driver.

For the record, ZeroMq is a very lightweight distributed messaging system
specially designed for high throughput/low latency scenarios.
Addition of support for ZeroMQ would help cinder scale out with high
performance and be highly available as there is no centralised broker.

DocImpact
Document the configurations for ZeroMQ driver for Cinder

Change-Id: Ic4b4301e5d7ca1692fc91155ba53f2dd12f99311
Closes-Bug: #1440631
partially Implements bp cinder-zeromq-support