Return updated volume type after updating
When updating volume type in db, it queries the object and
would like to return the updated object.

The problem is that the update session hasn't committed and
the update has not been implemented in db. And then the query is
executed in another session. As a result, the update function returns
an non-updated object.

The fix is to return updated object.

Closes-Bug: #1540190

Change-Id: I4b57ea69b1eebd62de707ca15aad32c53fca9f1d