HNAS driver: retry on Connection reset fix
When HNAS driver is not using the SSH option and there are lots of exports in
the backend, a exception of "Connection reset" is eventually raised in the
driver initialization.
This patch makes the HNAS driver retries when this exception is raised.

Change-Id: Ia6306d5d2d24556135505ddc33174f57c86d8555
Closes-Bug: 1541941
(cherry picked from commit 15a2752d512d7f74b9286e780428b559fe6582dd)