vzstorage: add ploop volume format support
This patch adds support for a volume format, which can
be used for containers in Virtuozzo Server.

This format is different from qcow2 images, because it consists
of 2 or more files in a directory.

The first file is an xml description of a bundle. It includes
paths to base image and deltas, which usually stored in the same
directory. Here is a description of base/delta images format -
https://openvz.org/Ploop/format.

Format name is a little ambiguous, because ploop is not only an
image format name, but also a technology, which allows to use
images, stored in files as block devices (like loop devices or
qemu-nbd), but it supports snapshots and has nearly native
performance. qemu-img has limited support of this format - it
can only convert image files, the format is called 'parallels'
in qemu-img.

The main utility to manage such disks is 'ploop'. Here is a man
page - https://static.openvz.org/vz-man/man8/ploop.8.gz.html.
It can create images, make snapshots and create ploop devices,
based on images.

This patch series makes possible all volume operations except for
snapshotting of 'in-use' volumes.

DocImpact
Depends-On: Iebd20544102672d7b7ca820c9e9005833ec2580e
Change-Id: Id3c902c38192fddc02529afe5af7abc213c7b57c