Remove Cisco FC Zone Manager Driver
After many attempts to get a Cisco CI running to validate the Cisco FCZM
driver over a few releases there is still no official third party CI
reporting. This was covered for some time due to the support of it being
covered by another CI, but that is no longer the case.

We can't include drivers that do not have test coverage and no report of
whether it is working as expected. Until there is test coverage, we will
need to remove this driver due to lack of CI.

DocImpact: Remove Cisci FCZM documentation

Change-Id: Idb7dfea74ce248d8f1209c55b62412be4f2e16e9