Provide user friendly message for FK failure
'cinder-manage db purge' command fails with DBReferenceError due
to FK constraint failure and exits with stack-trace on the command
prompt.

Made changes to give user friendly message to the user as well as
log appropriate error message in cinder-manage logs instead of
stack-trace and exit with non-zero code to notify user about
the failure.

Closes-Bug: #1542169
Change-Id: I24476af70b45150b9f9e4e5d874f46fd4c07e680