cinder volume HA
commit 2e22f0a3e6e64faef8d457060961e46a7ca46a9 removed the coupling
between volume and and its host. So, volume requests were sent to the
active volume service. We are running volume services in
active-passive mode with a VIP associated with the active one.
The volume service should listen on this VIP.
So, set the volume rpc host to this VIP.

Note - This requires a new config option 'rpc_zmq_vol_vip'
       to be set with volume vip value. This option can be removed
       after we move to active active configuration.

Testing done in DevTest env:
* Tested the patch with volume, scheduler and api services.
* Tested that the scheduler was able to send volume create requests to
  volume service
* Tested that the api was able to send delete volume delete requests to
  volume service

Closes-Bug: #JBS-105