Refactor sqlalchemy service methods
This patch refactors the sqlalchemy DB methods related to the service to
consolidate the methods into the minimum number that is actually needed,
to reduce the number of DB calls per operation, and to reduce the data
retrieved from the DB server.

For method consolidation the patch removes
`service_get_by_host_and_topic`, `service_get_by_args`,
`service_get_all_by_topic`, and `service_get_all_by_binary`, and
includes the functionality provided by those methods in `service_get`
and `service_get_all` methods.

To reduce the number of DB calls we won't retrieve the service from the
DB when deleting or updating a service.

And to reduce the data retrieved from the DB we filter UP nodes in the
DB instead of locally.

This patch is related to the efforts to support HA A/A in c-vol nodes.
Job distribution patches are dependent on this one.

Specs: https://review.openstack.org/327283
Implements: blueprint cinder-volume-active-active-support
Change-Id: I5c453dcb5c38301721c3017ba8e782c0fdf850e6