"volume service is down" warnings misinterpreted
We run cinder-volume service in active passive mode. There were "volume
service is down" warning messages in cinder-volume logs about the
passive volume service. This led to misinterpretation that there were
problems with the volume service.

So, this patch decreased the log level of this message to 'DEBUG'.
We should switch back once we have cinder-volume service running in
active-active mode as this message is a useful health check indicator.

Closes-Bug: #JBS-108