Adding general notes about rolling upgrades
There are some known problems that may happen during rolling (live)
upgrade. Even though feature is considered a tech-preview there's no
harm in communicating them in the release notes.

This commits adds also a note on the fact that now Cinder services
detect lowest RPC version and that it is needed to keep the `services`
table up to date.

Change-Id: Ic4c9188fb0bbd375d809c7dedd9e07df049ae648
Related-Blueprint: rpc-object-compatibility