Update quota_utils with import for keystone_auth
Currently, keystone_auth config group is only getting defined in
the cinder-api service. In order to be able to use the quota_utils
helper from both the c-api and c-vol, we need to make sure that
the CONF.keystone_authtoken always gets imported. This is needed
to properly handle nested quota deletes which occur in the c-vol
service.

Change-Id: I3ebb9979d37f62db161e2929fe2cff5a8ba2bf96