DRBD: Policy-based waiting for completion
The more nodes are in a cluster, the higher the chance that one or more
of them are not available. Waiting on all nodes will not work any more,
so give the user a way to define what conditions are acceptable (to the
management, etc.) to continue.

The upcoming DRBDmanage 1.0 release will have a (sample) policy plugin,
so make that available in Cinder, too.
Documentation is at
    http://drbd.linbit.com/users-guide-9.0/s-drbdmanage-deployment-policy.html

DocImpact
Change-Id: Ib2bc888ca5df87a4de97ee07601d0280a04c7a81