Pass RBD order to clone call
For cloning of a RBD the rbd_store_chunk_size information from
the cinder.conf should be used to calculate and pass the correct
order information to the clone() call of the rbd library.

Added new test to check for correctly from rbd_store_chunk_size
calculated order while cloning.

Closes-Bug: #1489134

Change-Id: Ic5714d3e0d6961bce6ff588006661618130dca07
Co-Authored-By: Logan V <logan2211@gmail.com>
Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>