Move replication_status update to init_with_rpc
We were using host_init to read replication_status from the
driver and update the service entry in the DB.  It turns out
that on a fresh install this doesn't actually work, because
while we have multiple init methods for the backend, the
Service entry isn't actually created in a fresh deploy until
AFTER init_host.  The result was that in some cases we were
trying to update a column on a non-existent Service in the
DataBase.

This patch moves the replication_status updates for the
service into the init_with_rpc method.  That method was
just a noop stub in the parent manager class, so we just
implement it in cinder.volume.manager and do what we need
with the replication update info.

Change-Id: I18b2658e2f93959f74377ccb86ce8b01b6970c60
Closes-Bug: #1555370