SMBFS: fix parsing volume type extra specs and metadata
When creating a new volume, the SMBFS volume driver can accept the
volume backing image file format via metadata or extra specs.

The issue is that the volume object the driver receives can be
either a SQLAlchemy model object, either a versioned object.

While there was an effort to keep versioned objects backwards
compatible with the SqlA objects, there still are some
inconsistencies regarding the volume type extra specs.

For this reason, the SMBFS driver fail to parse volume type extra
specs in case of versioned objects.

This patch fixes this issue, by making the driver able to parse
versioned objects as well.

Change-Id: I0385a5ba290ab03b676145d35206b1ed12899473
Closes-Bug: #1529877