3PAR fix create_cloned_volume for larger size
This patch fixes the 3PAR driver's create_cloned_volume
call in the case where a larger size for the volume was
specified.  When the new volume size is the same as the
source volume, we still use an 'online' copy, which is
3PAR speak for background copy.  When the new volume size
is larger than the source volume, we have to do an offline
copy, which entails creating a fresh volume with the requested
size, and then copying the bits from the old volume to the new.
This copy happens on the 3PAR itself and can take some time to
complete.

This patch also updates the minimum client version to 4.2.0

Change-Id: Ie2da441b32017f38e743189060209084e1f20974
Closes-bug: #1554740