Calculate virtual free capacity and notify
It calculates virtual free capacity for
pool and backend which reports definite
values. It will also notify various
capacity info to Ceilometer service.
The notification only occurs when
volume status is updated. It will give
users a bit knowledge about the current
storage usage.

As there are multiple schedulers and volumes, the
patch contains the logic only to send changed
capacity info to ceilometer as possible as. It adds
a new rpc call named notify_service_capabilities()
which fanout=false in volume manager. This means
that only one scheduler will receive it and send
the notification.

Co-Authored-By: Gorka Eguileor <geguileo@redhat.com>
Co-Authored-By: lisali <xiaoyan.li@intel.com>

DocImpact
Implements: blueprint capacity-headroom
Change-Id: Iff033d1b591fb3d9c0a5b9732c0c17d96ddbf712