Storwize: Multiple management IP not raising exceptions
In the case where both the san_ip and the storwize_san_secondary_ip
are both enabled and both fail to execute a ssh command, no
exceptions are raised.

This patch fixes this issue by creating a helper function to switch
the IP that _run_ssh is using and moving the code into a try/except
block.

Unit tests were also added to make sure all exceptions are covered.

Closes-Bug: 1558729
Change-Id: Ic472936046bab3a7c428a05bdcdc97492e52eb38