HPE LeftHand: extend volume if larger than snapshot
Extend volume if during create_volume_from_snapshot() the new volume
is larger than the snapshot it was created from. Added new test case.

Bump driver version to 2.0.11 and updated changelog entry.

Closes-Bug: #1560654

The volume has not been created with OVO and a seperate patch will be
created to switch the dictionary to OVO implementation regarding
volume / snapshot to follow up the OVO comments.

Co-Authored-By: Xinli Guan <xinli@us.ibm.com>

Change-Id: Iaeb2596ecd75b86a7a0d3da7fad0bf91b2a02fa0
Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>