image: don't use is_public with glance V2 API
Glance v2 API has replaced "is_public" argument with "visibility", and,
unlike "is_public" the default value of "visibility" returns all images,
which is what we want to get when no value has been provided.

This patch makes sure that we only send a default value for "is_public"
when it's not defined when working with API v1.

Closes-bug: #1201266
Co-Authored-By: Gorka Eguileor <geguileo@redhat.com>
Change-Id: I9ca13567744dbe5895012429679e013b859dbb03