Dont query image metadata for empty volume list
This query creates a table scan in sqlalchemy when there are no entries
in the IN() clause.

With a large image metadata table, this makes a volume list on a project
with no volumes take an incredible amount of time.

Change-Id: I51733e1e3c526f4ca46e2826c242924407fabbe4
Closes-Bug: #1561174
(cherry picked from commit 4c810374e399bac7019da6a8790f73017b06e11a)