Check hasattr before calling get in SolidFire
The addition of getattr to pull methods from other objects
is handy, but it's rather annoying during development if you
have something wrong and get the "recursion limit exceeded"
stack trace taking up the whole screen when you forget to
implement something.

This patch just adds a hasattr check before calling getattr.
If the check fails, we raise a NotImplemented, same result, but
much easier to look at and debug.

Change-Id: I4d657cea7b10ebd441ebf715c59b139904bc54b6