Storwize: fix unmapped multi-attached volume failure
On terminate_connection if the volume is not mapped to the host
entry in the connector but it is mapped to another host entry, then
continue on and don't try to remove the entry that we have already
determined does not exist. However return the known host name in when
this happens so host entry can be deleted and and ITM generated if
there are no other volumes mapped to that host entry for this connector.
When no mapping is found on for a volume return the host name which
triggers the deletion of the host entry and generation of the ITM if
there are no other volumes mapped to that host entry.

UT added.

Closes-Bug: #1563556

Change-Id: Ic0d618cadb5d32486113cc1767b7d1a8188d9d05