Implement replicate_volume & failover on SolidFire
This adds the Cheesecake Replication implementation
for the SolidFire driver.

We add the necessary methods for creating a replicated volume and
performing a failover properly that updates the status of volumes
on the failover host.

We also make sure that things like delete, attach etc work in a
failed over scenario.

For SolidFire we are using the remote cluster mvip as the
backend_id, here's an example replication_device entry for
a cinder.conf:
    replication_device =
    backend_id:192.168.139.102,mvip:192.168.139.102,login:admin,password:admin

Note that there's an independent fix from this patch series that needs
to be applied in order for things to work properly:
  https://review.openstack.org/#/c/299723/

The last remaining task is to ensure we detect and deal with a restart
when we're in a failed over state.

Change-Id: I78244da226abce2544d0c3196fccdec50bf5b8e5