RBD: delete snapshots if missing in the backend
snapshot delete should be more robust to also handle backend errors.
In case the backend does not have the image snapshot anymore,
log an info message and succeed the operation.
A first attempt to fix this problem was found on
https://review.openstack.org/#/c/151257/
This patch is already merged but it does not fix the problem
completely.

Change-Id: I1e80c2f71b79a5d40a0b7a42f03eb5bf911c24f0
Closes-Bug: #1415905